"Added new beanstalkd operation and response classes"
This commit added several new operation and response classes to the beanstalkd client. This includes operations like BuryOperation, DeleteOperation, KickOperation, ListTubesOperation etc and response like FoundResponse, KickedResponse, OkResponse etc.

These classes are necessary to handle different commands and responses from the Beanstalkd work queue. They simplify the process of interacting with the queue by abstracting the details of command and response handling. The classes follow a consistent structure which makes the client more maintainable and easier to understand. Several changes are also made in existing classes to align with these new classes.

The priority and delay parameters in put() and release() methods in DefaultClient class are now of Kotlin Duration type instead of UInt for better readability and understanding of the code. Modifications are made in some response classes to handle the responses correctly. UnitResponses are added to handle common responses. Changes in some other classes are also made for the better handelling of various cases. Some TODO implementations are now implemented. This commit makes the beanstalkd client more comprehensive and robust.