Operation now implements the Deferred<T> interface to combine the Operation with its result and make it awaitable.
Modified DefaultConnection to use a buffered Channel to execute operations in the order handle is called.

To not change the interface exposed to all Operation implementations, AbstractOperations still returns Result<T> from readResponse function and completes the deferred with this result.

BeanstalkdTaskIterator has been changed to use "reserveWithTimeout" with a 10s timeout to not block the use of ack and nack functions possibly forever. Combined with the channel implementation of DefaultConnection, this results in ack and nack calls being queued up until the timeout expired or a job has been reserved. At this point all pending operations are executed before a new reserve operation is issued.

Worker has been updated to handle a certain amount of parallel task with a specific amount of threads. This is handled using a channel of a specific size defined by WorkerConfig.parallelism. this ensures hat always just a specific amount of tasks is in flight. channel.suspend will suspend the iteration over the task queue until at least one space is available in the channel. Even though a custom dispatcher has been created with a fixed number of threads, it might be possible, that a substantial number of tasks are added to the processing in flight without any restrictions or guarantee for available processing resources.