Implement Cabinet API models and update Tag models
Expanded the functionality of the Mayan client by including the CabinetClient and updating Tag models. New data fields are introduced for the Cabinet model based on the ApiCabinet. The CabinetClient allows to better operate with the cabinets. Further, the MayanClient is improved by introducing lazy initialization for the new CabinetClient. In TagClient, a function for listing Tags is added. Basic CRUD operations are provided for the Tag model. This improvement enhances the interactability with Mayan's Cabinet and Tag APIs improving user experience and development speed.